+++ CHANGES.md
@@ -6,6 +6,7 @@
 - CI: Added support for Python 3.13
 - DMS: Fixed handling of primary keys
 - DMS: Added support for the previously missing `drop-table` operation
+- DMS: Fixed primary key / where clause rendering
 
 ## 2024/10/28 v0.0.22
 - DynamoDB/Testing: Use CrateDB nightly again